// enums/rssLinks.ts
export enum RSSlinks {
    TECH = 'https://feeds.bbci.co.uk/news/technology/rss.xml',
    SPORT = 'https://feeds.bbci.co.uk/news/sport/rss.xml',
    WORLD = 'https://feeds.bbci.co.uk/news/world/rss.xml',
    BUSINESS = 'https://feeds.bbci.co.uk/news/business/rss.xml',
    POLITICS = 'https://feeds.bbci.co.uk/news/politics/rss.xml',
    HEALTH = 'https://feeds.bbci.co.uk/news/health/rss.xml',
    SCIENCE = 'https://feeds.bbci.co.uk/news/science_and_environment/rss.xml',
    ENTERTAINMENT = 'https://feeds.bbci.co.uk/news/entertainment_and_arts/rss.xml',
    UK = 'https://feeds.bbci.co.uk/news/uk/rss.xml',
    ENGLAND = 'https://feeds.bbci.co.uk/news/england/rss.xml',
    WALES = 'https://feeds.bbci.co.uk/news/wales/rss.xml',
    SCOTLAND = 'https://feeds.bbci.co.uk/news/scotland/rss.xml',
    NORTHERN_IRELAND = 'https://feeds.bbci.co.uk/news/northern_ireland/rss.xml',
    MAGAZINE = 'https://feeds.bbci.co.uk/news/magazine/rss.xml',
    EDUCATION = 'https://feeds.bbci.co.uk/news/education/rss.xml',
    IN_PICTURES = 'https://feeds.bbci.co.uk/news/in_pictures/rss.xml',
    VIDEO_AND_AUDIO = 'https://feeds.bbci.co.uk/news/video_and_audio/rss.xml',
    NEWSBEAT = 'https://feeds.bbci.co.uk/news/newsbeat/rss.xml',
    STORIES = 'https://feeds.bbci.co.uk/news/stories/rss.xml',
    HAVE_YOUR_SAY = 'https://feeds.bbci.co.uk/news/have_your_say/rss.xml',
    WORLD_AFRICA = 'https://feeds.bbci.co.uk/news/world/africa/rss.xml',
    WORLD_ASIA = 'https://feeds.bbci.co.uk/news/world/asia/rss.xml',
    WORLD_EUROPE = 'https://feeds.bbci.co.uk/news/world/europe/rss.xml',
    WORLD_LATIN_AMERICA = 'https://feeds.bbci.co.uk/news/world/latin_america/rss.xml',
    WORLD_MIDDLE_EAST = 'https://feeds.bbci.co.uk/news/world/middle_east/rss.xml',
    WORLD_US_CANADA = 'https://feeds.bbci.co.uk/news/world/us_and_canada/rss.xml',
    CLIMATE = 'https://feeds.bbci.co.uk/news/science_and_environment/climate_change/rss.xml',
    CORONAVIRUS = 'https://feeds.bbci.co.uk/news/coronavirus/rss.xml'
}